refactor: kafka 관련 클래스 기능별 패키지 구조화
