+++ order-service/src/main/java/com/node5/orderservice/order/infrastructure/kafka/producer/StockRestoreProducer.java
@@ -1,6 +1,7 @@
-package com.node5.orderservice.order.infrastructure.kafka;
+package com.node5.orderservice.order.infrastructure.kafka.producer;
 
 import com.node5.common.event.StockRestoreEvent;
+import com.node5.orderservice.order.infrastructure.kafka.dto.StockRestoreKafkaRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
